have only untestable code in main.rs
